tree-construction/noscript01.dat

#data
<head><noscript><!doctype html><!--foo--></noscript>
#script-off
#errors
Line: 1 Col: 6 Unexpected start tag (head). Expected DOCTYPE.
Line: 1 Col: 31 Unexpected DOCTYPE. Ignored.
#new-errors
(1:1) missing-doctype
(1:17-1:32) misplaced-doctype
#document
| <html>
|   <head>
|     <noscript>
|       <!-- foo -->
|   <body>

#data
<head><noscript><html class="foo"><!--foo--></noscript>
#script-off
#errors
Line: 1 Col: 6 Unexpected start tag (head). Expected DOCTYPE.
Line: 1 Col: 34 html needs to be the first start tag.
#new-errors
(1:1) missing-doctype
#document
| <html>
|   class="foo"
|   <head>
|     <noscript>
|       <!-- foo -->
|   <body>

#data
<head><noscript></noscript>
#script-off
#errors
(1,6): expected-doctype-but-got-tag
#new-errors
(1:1) missing-doctype
#document
| <html>
|   <head>
|     <noscript>
|   <body>

#data
<head><noscript>   </noscript>
#script-off
#errors
Line: 1 Col: 6 Unexpected start tag (head). Expected DOCTYPE.
#new-errors
(1:1) missing-doctype
#document
| <html>
|   <head>
|     <noscript>
|       "   "
|   <body>

#data
<head><noscript><!--foo--></noscript>
#script-off
#errors
(1,6): expected-doctype-but-got-tag
#new-errors
(1:1) missing-doctype
#document
| <html>
|   <head>
|     <noscript>
|       <!-- foo -->
|   <body>

#data
<head><noscript><basefont><!--foo--></noscript>
#script-off
#errors
Line: 1 Col: 6 Unexpected start tag (head). Expected DOCTYPE.
#new-errors
(1:1) missing-doctype
#document
| <html>
|   <head>
|     <noscript>
|       <basefont>
|       <!-- foo -->
|   <body>

#data
<head><noscript><bgsound><!--foo--></noscript>
#script-off
#errors
Line: 1 Col: 6 Unexpected start tag (head). Expected DOCTYPE.
#new-errors
(1:1) missing-doctype
#document
| <html>
|   <head>
|     <noscript>
|       <bgsound>
|       <!-- foo -->
|   <body>

#data
<head><noscript><link><!--foo--></noscript>
#script-off
#errors
Line: 1 Col: 6 Unexpected start tag (head). Expected DOCTYPE.
#new-errors
(1:1) missing-doctype
#document
| <html>
|   <head>
|     <noscript>
|       <link>
|       <!-- foo -->
|   <body>

#data
<head><noscript><meta><!--foo--></noscript>
#script-off
#errors
Line: 1 Col: 6 Unexpected start tag (head). Expected DOCTYPE.
#new-errors
(1:1) missing-doctype
#document
| <html>
|   <head>
|     <noscript>
|       <meta>
|       <!-- foo -->
|   <body>

#data
<head><noscript><noframes>XXX</noscript></noframes></noscript>
#script-off
#errors
Line: 1 Col: 6 Unexpected start tag (head). Expected DOCTYPE.
#new-errors
(1:1) missing-doctype
#document
| <html>
|   <head>
|     <noscript>
|       <noframes>
|         "XXX</noscript>"
|   <body>

#data
<head><noscript><style>XXX</style></noscript>
#script-off
#errors
Line: 1 Col: 6 Unexpected start tag (head). Expected DOCTYPE.
#new-errors
(1:1) missing-doctype
#document
| <html>
|   <head>
|     <noscript>
|       <style>
|         "XXX"
|   <body>

#data
<head><noscript></br><!--foo--></noscript>
#script-off
#errors
Line: 1 Col: 6 Unexpected start tag (head). Expected DOCTYPE.
Line: 1 Col: 21 Element br not allowed in a inhead-noscript context
Line: 1 Col: 21 Unexpected end tag (br). Treated as br element.
Line: 1 Col: 42 Unexpected end tag (noscript). Ignored.
#new-errors
(1:1) missing-doctype
(1:17-1:22) disallowed-content-in-noscript
#document
| <html>
|   <head>
|     <noscript>
|   <body>
|     <br>
|     <!-- foo -->

#data
<head><noscript><head class="foo"><!--foo--></noscript>
#script-off
#errors
Line: 1 Col: 6 Unexpected start tag (head). Expected DOCTYPE.
Line: 1 Col: 34 Unexpected start tag (head).
#new-errors
(1:1) missing-doctype
#document
| <html>
|   <head>
|     <noscript>
|       <!-- foo -->
|   <body>

#data
<head><noscript><noscript class="foo"><!--foo--></noscript>
#script-off
#errors
Line: 1 Col: 6 Unexpected start tag (head). Expected DOCTYPE.
Line: 1 Col: 34 Unexpected start tag (noscript).
#new-errors
(1:1) missing-doctype
#document
| <html>
|   <head>
|     <noscript>
|       <!-- foo -->
|   <body>

#data
<head><noscript></p><!--foo--></noscript>
#script-off
#errors
Line: 1 Col: 6 Unexpected start tag (head). Expected DOCTYPE.
Line: 1 Col: 20 Unexpected end tag (p). Ignored.
#new-errors
(1:1) missing-doctype
(1:17-1:21) end-tag-without-matching-open-element
#document
| <html>
|   <head>
|     <noscript>
|       <!-- foo -->
|   <body>

#data
<head><noscript><p><!--foo--></noscript>
#script-off
#errors
Line: 1 Col: 6 Unexpected start tag (head). Expected DOCTYPE.
Line: 1 Col: 19 Element p not allowed in a inhead-noscript context
Line: 1 Col: 40 Unexpected end tag (noscript). Ignored.
#new-errors
(1:1) missing-doctype
(1:17-1:20) disallowed-content-in-noscript
#document
| <html>
|   <head>
|     <noscript>
|   <body>
|     <p>
|       <!-- foo -->

#data
<head><noscript>XXX<!--foo--></noscript></head>
#script-off
#errors
Line: 1 Col: 6 Unexpected start tag (head). Expected DOCTYPE.
Line: 1 Col: 19 Unexpected non-space character. Expected inhead-noscript content
Line: 1 Col: 30 Unexpected end tag (noscript). Ignored.
Line: 1 Col: 37 Unexpected end tag (head). Ignored.
#new-errors
(1:1) missing-doctype
(1:17-1:20) disallowed-content-in-noscript
#document
| <html>
|   <head>
|     <noscript>
|   <body>
|     "XXX"
|     <!-- foo -->

#data
<head><noscript>
#script-off
#errors
(1,6): expected-doctype-but-got-tag
(1,6): eof-in-head-noscript
#new-errors
(1:1) missing-doctype
(1:17) open-elements-left-after-eof
#document
| <html>
|   <head>
|     <noscript>
|   <body>
